requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-core==1.34.0
azure-identity==1.21.0
backoff==2.2.1
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
cloudevents==1.11.0
colorama==0.4.6
cryptography==44.0.3
dapr==1.15.0
dapr-agents==0.8.2
dapr-ext-fastapi==1.15.0
dapr-ext-workflow==1.15.0
deprecation==2.1.0
distlib==0.3.9
distro==1.9.0
durabletask-dapr==0.2.0a7
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
identify==2.6.10
idna==3.10
jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.13.1
mdurl==0.1.2
msal==1.32.3
msal-extensions==1.3.1
multidict==6.4.3
nodeenv==1.9.1
numpy==2.2.2
openai==1.75.0
openapi-pydantic==0.5.1
packaging==25.0
platformdirs==4.3.7
posthog==5.4.0
pre-commit==4.2.0
prompty==0.1.50
propcache==0.3.1
protobuf==5.29.4
pycparser==2.22
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.10.1
pygments==2.19.1
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.5
rich==13.9.4
rpds-py==0.25.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.46.2
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
virtualenv==20.31.1
yarl==1.20.0
